{{cookiecutter.repo_name}}/requirements/base.txt
# Django 1.7 beta release
https://github.com/django/django/archive/1.7c1.zip

django-configurations==0.8
django-extensions==1.3.8
django-secure==1.0

# images
Pillow==2.4.0

# Postgres client driver
psycopg2==2.5.3

structlog==0.4.1
raven==5.0.0

# caching backend
django-redis-cache==0.13.0

uwsgi==2.0.6
